Extract more utils into the common module
